Add existing file

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,23 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+    
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+    
     /**
      * Display a listing of the resource.
      *
@@ -42,14 +54,36 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' =>'required'
+            'body' =>'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+        
+        //file upload
+        if($request->hasFile('cover_image')){
+            //get file name with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            
+            //get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            
+            //file name to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
         
         //create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         
         return redirect('/posts')->with('success', 'Post Created');
@@ -76,6 +110,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        
+        //check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
+        
         return view('posts.edit')->with('post',$post);
     }
 
@@ -93,10 +133,31 @@ class PostsController extends Controller
             'body' =>'required'
         ]);
         
+        //file upload
+        if($request->hasFile('cover_image')){
+            //get file name with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            
+            //get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            
+            //file name to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+        
         //create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         
         return redirect('/posts')->with('success', 'Post Updated');
@@ -111,6 +172,16 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        
+        //check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
+        
+        if($post->cover_image != 'noimage.jpg'){
+            //dekete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         
         return redirect('/posts')->with('success', 'Post Deleted');
